+++ R/overlay_sdm.R
@@ -27,7 +27,13 @@
 #'   geometry. Note that this means all columns of \code{sdm} not in
 #'   \code{sdm.idx} will not be in the returned object.
 #'   Because the data are considered spatially intensive, the \code{agr}
-#'   attribute will be set as 'constant' for all columns in the returned object
+#'   attribute will be set as 'constant' for all columns in the returned object.
+#'
+#'   Additionally, the output will match the class of \code{sdm}, with regards
+#'   to the classes tbl_df, tbl, and data.frame. This means that, in addition to
+#'   being an \code{sf} object, if \code{sdm} is a tibble then the output will
+#'   also be a tibble, while if \code{sdm} is just a data frame then the output
+#'   will not be a tibble.
 #'
 #' @references Goodchild, M.F. & Lam, N.S.-N. (1980) Areal interpolation:
 #'   a variant of the traditional spatial problem. Geo-Processing, 1, 297-312.
@@ -64,6 +70,11 @@ overlay_sdm <- function(base.geom, sdm, sdm.idx, overlap.perc) {
     stop("'base.geom' must be of class 'sfc'")
   }
   if (!inherits(sdm, "sf")) stop("'sdm' must be of class 'sf'")
+  if (!all(class(sdm) %in% c("tbl_df", "tbl", "data.frame", "sf")))
+    warning("'sdm' contains classes other than tbl_df, tbl, data.frame, or sf. ",
+            "These classes will be dropped, and not included in the output sf object")
+
+
   if (!(is.numeric(overlap.perc) & between(overlap.perc, 0, 100))) {
     stop("'overlap.perc' must be a number between 0 and 100 (inclusive)")
   }
@@ -261,6 +272,11 @@ overlay_sdm <- function(base.geom, sdm, sdm.idx, overlap.perc) {
 
   #----------------------------------------------------------------------------
   # 5) Create sf obj w/predicted densities and base geom to make overlaid SDM
+
+  # Ensure that output sdm has the same class as input sdm.
+  # Specifically, drop "tbl_df" and "tbl" if necessary.
+  if (!inherits(sdm, c("tbl_df"))) new.dens.df <- as.data.frame(new.dens.df)
+
   stopifnot(
     nrow(new.dens.df) == nrow(base.geom),
     inherits(new.dens.df, "data.frame")